Implement inotify to monitor and handle move and deletion

+++ src/Window.zig
@@ -29,7 +29,7 @@ pub const Event = union(enum) {
     },
     resize: struct { c_int, c_int },
     image: struct {
-        index: usize,
+        index: i32,
         image: Image,
     },
 };
